Bio pool freeing
When you allocate a bio from a bio pool, to free it you have to know
where it came from; this adds a flag which, if set, means bi_destructor
is the pointer to the pool and bio_put() can do the right thing.

This is used in bcache, so we can cleanly use per device bio pools.

Change-Id: I41c718265a9f7a87e5aa804abb07ff6d38a7239e