bcache: Auxiliary binary search tree for bsets
Recent profiling has showed that bcache's CPU usage is dominated by the
binary searches within a btree node.

This is unsurprising, those binary searches look like random memory
accesses and with a btree that might be 100 mb they're usually not going
to be in cache.

So to speed this up we construct a partial binary search tree in an
array. By using a binary tree in an array our memory access patterns are
as good as they can be - the children of any node are adjacent to each
other in memory.

We construct the binary search trees after btree_sort() finishes - this
means we only have to pay the cost of building them when the set isn't
going to change for awhile.

And they don't have to be very big and cost us much memory since they
don't need to eliminate the binary search, just reduce the amount of
work it has to do.

Change-Id: I5cc90ef6b6787457fb8bee83c2a0b1c6e6993927
Signed-off-by: Kent Overstreet <koverstreet@google.com>