bcache: Compressed auxiliary search tree
Previously, the auxiliary binary search tree code used full 16 byte
keys and a separate array of ints for indexing into the sorted set.

It'd be nice if we could make the auxiliary search tree big enough to
get us down to one cacheline in the set, but with the full keys that'd
take way too much memory.

But while key has ~80 bits we sort on, sets have keys that are mostly
adjacent and so don't differ in all that many of those 80 bits.

So for the auxiliary search tree we now use basically a floating point
key, and then to index into the array store the different between (l+m)/2.
This lets us fit the whole key and pointer into 32 bits - see struct
bkey_float.

Change-Id: I05e1ae54c1b8ba51d97b5a7db90df1d1c9f8caa4
Signed-off-by: Kent Overstreet <koverstreet@google.com>