bcache: Seed pointer into btree node crc
Previously, the checksum for btree nodes depended only on the contents
of the node. This means that the checksum verifies the data, but it
doesn't verify that it's the data we want.

This change seeds the pointer (which includes the gen) into the
checksum, so if when reading a btree node we get a valid node but not
the node we wanted, the checksum will fail.

Change-Id: I60c33d946a92f26b50eb22b01d16fed28df98003
Signed-off-by: Kent Overstreet <koverstreet@google.com>