bcache: A trivial locking optimization
For writes, we need to make sure we don't do a writethrough write that
overlaps with background writeback of dirty data. We do this by
checking the rb tree that background writeback uses.

But that requires taking a spinlock and walking the rb tree for every
write bio we process. We're only doing background writeback to a small
region of the backing device at any given time, so now we keep track of
what that range is, and only check the rb tree if a bio is in that
range.

Change-Id: I58148d9e87439a479857292ae1aadc4a7c0274d8
Signed-off-by: Kent Overstreet <koverstreet@google.com>