bcache: Take out an old hacky optimization
Pre journalling, writes had to wait on btree writes, and we had to do a
lot more btree writes; thus it was worth going to a lot of effort to
optimize them.

We used to try to avoid a btree write if we were inserting a key with no
pointers and it didn't actually overwrite anything. This is just
needless complexity now, get rid of it.

Change-Id: Id4df38f1c5e741269be2b264db6d0c07d0fea8a9
Signed-off-by: Kent Overstreet <koverstreet@google.com>